Get or Set the Publisher of a Dataset Object

The publisher is the entity responsible for holding, archiving, releasing, or distributing the resource. It is typically included in dataset citation metadata.

For software, this might refer to a code repository (e.g., GitHub). If both a hosting platform and a producing institution are involved, use the publisher for the institution and creator() with contributorType = "hostingInstitution" for the platform.

Usage

publisher(x)

publisher(x, overwrite = TRUE) <- value

Arguments

x

A dataset object created with dataset_df() or as_dataset_df().

overwrite

Logical. Should existing publisher metadata be overwritten? Defaults to FALSE. If FALSE and the field exists, a warning is issued.

value

A character string specifying the publisher.

Value

A character string of length one containing the "publisher" attribute. When assigning, the updated object x is returned invisibly.

Details

Adds or retrieves the optional "publisher" attribute for a dataset object. This property aligns with dct:publisher (Dublin Core) and publisher (DataCite).

See also

Examples

publisher(orange_df) <- "Wiley"
publisher(orange_df)
#> [1] "Wiley"